refactor(worktree): split background pane exit-output recording out
The sequenced-command fix pushed launch-worktree-background-terminals.ts
to 303 effective lines, over the 300 cap. Moved persistExitedPaneOutput
and registerBackgroundPaneBuffer, which are one concern, into
background-pane-exit-output.ts. No behavior change: dropping the
registerEagerPtyBuffer call still reddens the background-terminals suite.
